Add Bandwidth concept
With these changes, the Bandwidth concept from the bills is
included. However, while accloudtant counts these separately for each
CSV file, this might not look correct, as there are traffic for some
services, to be cancelling the accountable traffic for another
services. For example, S3 traffic to CloudFront is free, so the billed
traffic for the latter should be reduced with some of the former.